backend completed

+++ flickart/src/main/java/com/flickart/model/Cart.java
@@ -36,7 +36,7 @@ public class Cart {
         return totalAmount;
     }
 
-    public void setTotalAmount(double totalAmount) {
+    public void setTotalAmount(float totalAmount) {
         this.totalAmount = totalAmount;
     }
