+++ flickart/src/main/java/com/flickart/model/CartItem.java
@@ -53,7 +53,7 @@ public class CartItem {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
 }
